Fix invalid label position attribute
`aria-label-position-` is not a valid aria prop. (I assume this was inteded to be `aria-label-position`)

```stacktrace
react-dom.development.js:86
Warning: Invalid aria prop `aria-label-position-` on <span> tag. For details, see https://reactjs.org/link/invalid-aria-props
    at span
    at O2 (plugin:obsidian-plugin-update-tracker:38766:6)
    at div
    at div
    at O2 (plugin:obsidian-plugin-update-tracker:38766:6)
    at div
    at DismissedPluginVersions (plugin:obsidian-plugin-update-tracker:41337:3)
    at DismissedPluginVersionsConnected (plugin:obsidian-plugin-update-tracker:41315:3)
    at Provider (plugin:obsidian-plugin-update-tracker:34816:10)
```

It also does not induce the intended behavior of placing the tooltip above the icon.

It seems this attribute (or at least, the intended `aria-label-position`) was used in a previous version of obsidian (not sure exactly when it changed), but currently (v1.7), obsidian uses `data-tooltip-position`

from obsidian's `app.js` (v1.7.5):

```javascript
function RS(e, t) {
    var n = t || {}
      , i = n.placement
      , r = n.classes
      , o = n.delay;
    i && "bottom" !== i && e.setAttribute("data-tooltip-position", i),
    r && e.setAttribute("data-tooltip-classes", r.join(" ")),
    o && e.setAttribute("data-tooltip-delay", String(o))
}
```